fix(ci): python-coverage-shards — split operator-directive tests so test file stays under the 1500-line maintainability limit
The #447 guide tests pushed test_pr_monitor_operator_hints.py to 1537 lines,
tripping test_first_party_code_files_stay_under_line_limit. Move the two
operator-directive tests into a focused test_pr_monitor_operator_directive.py.
